인덱스 수정, 위반 문구 추가, view 추가

+++ src/main/java/mapper/PostMapper.java
@@ -26,7 +26,7 @@ public interface PostMapper {
 	
 	int update(Post post);
 	
-	int increaseViewCount(Long pno);
+	int increaseViewCount(int pno);
 	
 	int delete(Long pno);
 }
